+++ Color/src/Graphics/Color/Model/X.hs
@@ -1,4 +1,5 @@
 {-# LANGUAGE BangPatterns #-}
+{-# LANGUAGE DataKinds #-}
 {-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE FlexibleInstances #-}
 {-# LANGUAGE GeneralizedNewtypeDeriving #-}
@@ -25,6 +26,7 @@ module Graphics.Color.Model.X
   , rgb2y
   ) where
 
+import Data.List.NonEmpty
 import Data.Coerce
 import Foreign.Storable
 import Graphics.Color.Model.Internal
@@ -65,6 +67,11 @@ instance Elevator e => Show (Color X e) where
 -- | `X` color model
 instance Elevator e => ColorModel X e where
   type Components X e = e
+  type ChannelCount X = 1
+  channelCount _ = 1
+  {-# INLINE channelCount #-}
+  channelNames _ = "Gray" :| []
+  channelColors _ = V3 0x80 0x80 0x80 :| []
   toComponents (X y) = y
   {-# INLINE toComponents #-}
   fromComponents = X