+++ Color/src/Graphics/Color/Space/CIE1931/RGB.hs
@@ -60,6 +60,11 @@ instance (Typeable l, Elevator e) => Show (Color (CIERGB l) e) where
 -- | `CIERGB` color space
 instance (Typeable l, Elevator e) => ColorModel (CIERGB l) e where
   type Components (CIERGB l) e = (e, e, e)
+  type ChannelCount (CIERGB l) = 3
+  channelCount _ = 3
+  {-# INLINE channelCount #-}
+  channelNames _ = channelNames (Proxy :: Proxy (Color CM.RGB e))
+  channelColors _ = channelColors (Proxy :: Proxy (Color CM.RGB e))
   toComponents = toComponents . unColorRGB
   {-# INLINE toComponents #-}
   fromComponents = mkColorRGB . fromComponents